requirements.txt
streamlit
pandas
numpy
nltk
scikit-learn
python-dotenv
flask